% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columns.R
\name{columns}
\alias{columns}
\title{Add columns to the Query}
\usage{
columns(theObject, string)
}
\arguments{
\item{theObject}{A Query class.}

\item{string}{What to show in columns.}
}
\value{
A Query.
}
\description{
Add columns to the Query
}
\examples{
library(magrittr)
Query() \%>\% columns(c("[Measures].[Internet Sales Count]", "[Measures].[Internet Sales-Sales Amount]"))
}
