% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\docType{methods}
\name{cube}
\alias{cube}
\alias{cube,Query-method}
\alias{cube,character,ANY-method}
\title{Add cube to the Query}
\usage{
cube(theObject, string)

\S4method{cube}{Query}(theObject, string)
}
\arguments{
\item{theObject}{A Query class.}

\item{string}{A name of the cube.}
}
\value{
A Query.
}
\description{
Add cube to the Query
}
\examples{
library(magrittr)
Query() \%>\% cube("[Adventure Works]")
}
