% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\docType{methods}
\name{execute}
\alias{execute}
\alias{execute,Query-method}
\alias{execute,character,ANY-method}
\title{Execute query}
\usage{
execute(theObject)

\S4method{execute}{Query}(theObject)
}
\arguments{
\item{theObject}{Query.}
}
\value{
A tibble.
}
\description{
Execute query
}
\examples{
\dontrun{
Sys.setenv(adomd_path = "/Program Files/Microsoft.NET/ADOMD.NET/160", 
conn_str = "Data Source=localhost; Provider=MSOLAP; 
initial catalog=Analysis Services Tutorial")
Query() \%>\% execute()
}
}
