% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute.R
\name{execute}
\alias{execute}
\title{Execute query}
\usage{
execute(theObject)
}
\arguments{
\item{theObject}{Query.}
}
\value{
A tibble.
}
\description{
Execute query
}
\examples{
\dontrun{
library(magrittr)
Sys.setenv(adomd_path = "/Program Files/Microsoft.NET/ADOMD.NET/160", conn_str = "Data Source=localhost; Provider=MSOLAP; initial catalog=Analysis Services Tutorial")
Query() \%>\% execute()
}
}
