% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\docType{methods}
\name{explore}
\alias{explore}
\alias{explore,Query-method}
\alias{explore,character,ANY-method}
\title{Explore query}
\usage{
explore(theObject)

\S4method{explore}{Query}(theObject)
}
\arguments{
\item{theObject}{Query.}
}
\value{
A string.
}
\description{
Explore query
}
\examples{
library(magrittr)
Query() \%>\% explore()
}
