% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore query}
\usage{
explore(theObject)
}
\arguments{
\item{theObject}{Query.}
}
\value{
A string.
}
\description{
Explore query
}
\examples{
library(magrittr)
Query() \%>\% explore()
}
