% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdx.R
\docType{methods}
\name{mdx}
\alias{mdx}
\alias{mdx,Query-method}
\alias{mdx,character,ANY-method}
\title{Optionally add mdx to the Query}
\usage{
mdx(theObject, string)

\S4method{mdx}{Query}(theObject, string)
}
\arguments{
\item{theObject}{A Query class.}

\item{string}{What to use as mdx.}
}
\value{
A Query.
}
\description{
Optionally add mdx to the Query
}
\examples{
library(magrittr)
Query() \%>\% mdx("SELECT {[Measures].[Internet Sales Count], 
[Measures].[InternetSales-Sales Amount]} ON COLUMNS, 
{[Product].[Product Line].[Product Line].MEMBERS} ON ROWS 
FROM [Analysis Services Tutorial] 
WHERE [Sales Territory].[Sales Territory Country].[Australia]")
}
