% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\docType{class}
\name{Query-class}
\alias{Query-class}
\alias{Query}
\title{Query Class}
\description{
Provides convenient & easy access to Ocean Networks Canada's data.
For detailed information and usage examples, visit our
\href{https://wiki.oceannetworks.ca/display/O2A/R+Client+Library}{official Documentation}.
}
\section{Slots}{

\describe{
\item{\code{cube}}{character. User token}

\item{\code{columns}}{logical. Print verbose debug comments}

\item{\code{rows}}{numeric. Number of seconds before a request to the API is canceled}

\item{\code{slicers}}{character. Base URL for API requests}

\item{\code{mdx}}{character. Output path for downloaded files}
}}

