% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{rows}
\alias{rows}
\title{Add rows to the Query}
\usage{
rows(theObject, string)
}
\arguments{
\item{theObject}{A Query class.}

\item{string}{What to show in rows.}
}
\value{
A Query.
}
\description{
Add rows to the Query
}
\examples{
library(magrittr)
Query() \%>\% rows(c("[Product].[Product Line].[Product Line].MEMBERS"))
}
