% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicers.R
\name{slicers}
\alias{slicers}
\title{Add slicers to the Query}
\usage{
slicers(theObject, string)
}
\arguments{
\item{theObject}{A Query class.}

\item{string}{What to use as slicers.}
}
\value{
A Query.
}
\description{
Add slicers to the Query
}
\examples{
library(magrittr)
Query() \%>\% slicers(c("[Sales Territory].[Sales Territory Country].[Australia]"))
}
